BasicTests: add ETC 'Atlantis' difficulty test
This was generated by 'piggy-backing' the adjacent difficultyByzantium.json
tests, and using multi-geth to run those same tests, but instead
of checking values (ie testing), the generated 'CurrentDifficulty'
values were instead written to each test and included in the suite.

These tests are intended to be run with an ETC 'Atlantis' (ECIP1054)
satisfying configuration.